Modelo Python/requirements.txt
absl-py==2.3.1
anyio==4.10.0
astunparse==1.6.3
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.1
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.32.0
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-bcb==0.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikeras==0.13.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yfinance==0.2.65
